core/users/views.py
from rest_framework import generics
from .models import User
from .serializers import UserSerializer


# Tạo user
class UserCreateView(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer


# Xoá user theo id
class UserDeleteView(generics.DestroyAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer
    lookup_field = 'id'

